requirements.txt
boto3
Jinja2
python-dateutil
nose
